// src/types/appointment.d.ts
type BookingItem = {
    id: string; // UUID for the booking item
    serviceName: string; // Name of the service
    date: string;
    memberName: string; // Name of the member
    serviceId: string | null; // ID of the service
    memberId: number | null; // ID of the member
    appointmentId: number; // ID of the associated appointment
    startTime: number; // Start time in seconds since midnight
    endTime: number; // End time in seconds since midnight
    price: number; // Price of the service
    duration: number; // Duration in seconds
    discountPrice: number; // Discounted price
    commissionFees: number; // Commission fees
};
// ...
export type AppointmentForAll = {
    id: string; // Appointment ID
    createdAt: string; // ISO date string for when the record was created
    updatedAt: string; // ISO date string for when the record was last updated
    deletedAt: string | null; // ISO date string or null for deleted records
    date: string; // Appointment date in ISO string format
    username: string; // Username of the customer
    notes: string | null; // Notes for the appointment
    status: "pending" | "confirmed" | "cancelled" | "completed"; // Status of the appointment
    phone: string; // Phone number of the customer
    email: string; // Email of the customer
    profilePicture: string; // URL of the customer's profile picture
    gender: "male" | "female" | "none"; // Gender of the customer
    totalTime: number; // Total duration in seconds
    totalPrice: number; // Total price for the appointment
    discountPrice: number; // Discounted total price
    isOnlineBooking: boolean; // Whether the booking is online
    startTime: number; // Start time in seconds since midnight
    endTime: number; // End time in seconds since midnight
    orgId: number; // Organization ID
    organization: Organization;
    bookingItems: BookingItem[]; // Array of booking items
    token: number;
};